requirements.txt
asgiref==3.5.0
click==8.1.0
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
Flask==2.1.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
Werkzeug==2.1.0
whitenoise==6.0.0
